:art: Refactor to utilize grpc-gateway template
