+++ highway/crud_channel.go
@@ -8,8 +8,8 @@ import (
 	ctv1 "github.com/sonr-io/blockchain/x/channel/types"
 	otv1 "github.com/sonr-io/blockchain/x/object/types"
 	"github.com/sonr-io/core/channel"
-	ct "go.buf.build/grpc/go/sonr-io/blockchain/channel"
-	v1 "go.buf.build/grpc/go/sonr-io/core/highway/v1"
+	ct "go.buf.build/sonr-io/grpc-gateway/sonr-io/blockchain/channel"
+	v1 "go.buf.build/sonr-io/grpc-gateway/sonr-io/core/highway/v1"
 )
 
 // CreateChannel creates a new channel.
