+++ highway/crud_object_test.go
@@ -6,8 +6,8 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/require"
-	ot "go.buf.build/grpc/go/sonr-io/blockchain/object"
-	"go.buf.build/grpc/go/sonr-io/blockchain/registry"
+	ot "go.buf.build/sonr-io/grpc-gateway/sonr-io/blockchain/object"
+	"go.buf.build/sonr-io/grpc-gateway/sonr-io/blockchain/registry"
 )
 
 func TestMsgCreateObject_ValidateBasic(t *testing.T) {
@@ -34,7 +34,7 @@ func TestMsgCreateObject_ValidateBasic(t *testing.T) {
 				InitialFields: []*ot.TypeField{
 					{
 						Name: "another label",
-						Kind:  0,
+						Kind: 0,
 					},
 				},
 				Session: &registry.Session{
