+++ host/exchange/protocol.go
@@ -10,10 +10,10 @@ import (
 	"github.com/sonr-io/core/device"
 	"github.com/sonr-io/core/host"
 	"github.com/sonr-io/core/motor/config"
-	v1 "go.buf.build/grpc/go/sonr-io/core/host/exchange/v1"
-	types "go.buf.build/grpc/go/sonr-io/core/types/v1"
+	v1 "go.buf.build/sonr-io/grpc-gateway/sonr-io/core/host/exchange/v1"
+	types "go.buf.build/sonr-io/grpc-gateway/sonr-io/core/types/v1"
 
-	motor "go.buf.build/grpc/go/sonr-io/core/motor/v1"
+	motor "go.buf.build/sonr-io/grpc-gateway/sonr-io/core/motor/v1"
 	"google.golang.org/protobuf/proto"
 )
