+++ host/transmit/config.go
@@ -8,8 +8,8 @@ import (
 
 	"github.com/sonr-io/core/device"
 
-	v1 "go.buf.build/grpc/go/sonr-io/core/host/transmit/v1"
-	types "go.buf.build/grpc/go/sonr-io/core/types/v1"
+	v1 "go.buf.build/sonr-io/grpc-gateway/sonr-io/core/host/transmit/v1"
+	types "go.buf.build/sonr-io/grpc-gateway/sonr-io/core/types/v1"
 )
 
 // Transfer Protocol ID's
