+++ host/transmit/item.go
@@ -8,9 +8,9 @@ import (
 
 	"github.com/libp2p/go-msgio"
 	"github.com/sonr-io/core/motor/config"
-	v1 "go.buf.build/grpc/go/sonr-io/core/host/transmit/v1"
-	motor "go.buf.build/grpc/go/sonr-io/core/motor/v1"
-	types "go.buf.build/grpc/go/sonr-io/core/types/v1"
+	v1 "go.buf.build/sonr-io/grpc-gateway/sonr-io/core/host/transmit/v1"
+	motor "go.buf.build/sonr-io/grpc-gateway/sonr-io/core/motor/v1"
+	types "go.buf.build/sonr-io/grpc-gateway/sonr-io/core/types/v1"
 )
 
 // ReadFromStream reads the item from the stream
