+++ host/transmit/session.go
@@ -6,10 +6,10 @@ import (
 	"github.com/libp2p/go-libp2p-core/network"
 	"github.com/libp2p/go-msgio"
 	"github.com/sonr-io/core/motor/config"
-	v1 "go.buf.build/grpc/go/sonr-io/core/host/transmit/v1"
-	motor "go.buf.build/grpc/go/sonr-io/core/motor/v1"
+	v1 "go.buf.build/sonr-io/grpc-gateway/sonr-io/core/host/transmit/v1"
+	motor "go.buf.build/sonr-io/grpc-gateway/sonr-io/core/motor/v1"
 
-	types "go.buf.build/grpc/go/sonr-io/core/types/v1"
+	types "go.buf.build/sonr-io/grpc-gateway/sonr-io/core/types/v1"
 )
 
 // NewInSession creates a new Session from the given payload with Incoming direction.
